requirements.txt
aiosmtplib==5.0.0
asyncpg==0.31.0
cryptography==46.0.3
fastapi==0.127.0
httpx==0.28.1
jinja2==3.1.6
jwcrypto==1.5.6
pydantic[email]==2.12.5
pydantic-settings==2.12.0
uvicorn[standard]==0.40.0
dnspython==2.7.0
